+++ test/test_tensors.cpp
@@ -5,14 +5,14 @@
 namespace taco {
 namespace test {
 
-std::vector<std::vector<ModeFormatPack>> generateModeTypes(size_t order) {
+std::vector<std::vector<ModeFormatPack>> generateModeTypes(int order) {
   taco_iassert(order > 0);
   std::vector<size_t> divisors(order);
 
   const size_t numModeTypes = 2;
 
   divisors[0] = 1;
-  for (size_t i = 1; i < order; ++i) {
+  for (int i = 1; i < order; ++i) {
     divisors[i] = numModeTypes * divisors[i - 1];
   }
   
@@ -21,7 +21,7 @@ std::vector<std::vector<ModeFormatPack>> generateModeTypes(size_t order) {
   std::vector<std::vector<ModeFormatPack>> levels(numPermutations);
   for (size_t i = 0; i < levels.size(); ++i) {
     std::vector<ModeFormatPack> level;
-    for (size_t j = 0; j < order; ++j) {
+    for (int j = 0; j < order; ++j) {
       switch ((i / divisors[j]) % numModeTypes) {
         case 0:
           level.push_back(Dense);
@@ -40,13 +40,13 @@ std::vector<std::vector<ModeFormatPack>> generateModeTypes(size_t order) {
   return levels;
 }
 
-std::vector<std::vector<size_t>> generateModeOrderings(size_t order) {
-  std::vector<size_t> modeOrdering(order);
-  for (size_t i = 0; i < order; ++i) {
+std::vector<std::vector<int>> generateModeOrderings(int order) {
+  std::vector<int> modeOrdering(order);
+  for (int i = 0; i < order; ++i) {
     modeOrdering[i] = i;
   }
 
-  std::vector<std::vector<size_t>> modeOrderings;
+  std::vector<std::vector<int>> modeOrderings;
   do {
     modeOrderings.push_back(modeOrdering);
   } while (std::next_permutation(modeOrdering.begin(), modeOrdering.end()));
